Further alignment of opcodes and word implementations between 1.6.x and master.
	modified:   src/ForthCompiler.cpp
	modified:   src/ForthCompiler.h
	modified:   src/ForthVM.cpp
	modified:   src/ForthVM.h
	modified:   src/ForthWords.h
	modified:   src/fbc.h
	modified:   src/kforth.cpp
	modified:   src/vm32-common.s

+++ src/ForthCompiler.h
@@ -86,7 +86,7 @@ public:
 
 
 bool IsForthWord (char*, WordListEntry*);
-int  ForthCompiler (vector<byte>*, int*);
+int  ForthCompiler (vector<byte>*, long int*);
 void CompileWord (WordListEntry);
 void OutputForthByteCode (vector<byte>*);
 void SetForthInputStream (istream&);
